03. TextExtractor_OCR(Deploy_AWS)/model_AWS/requirements.txt
opencv-python==4.1.0.25
pytesseract==0.3.0
pillow==5.1.0
Flask==1.0.2
gunicorn==19.9.0
pandas==0.25.1
numpy==1.16.4
